+++ doc/change-log.md
@@ -2,6 +2,7 @@
 ## Improvement
 - Added `StringUtils.strMaxLen()`
 - Added `RawReply.invalidParamRaw()`, `RawReply.TIMEOUT`, `Reply.TIMEOUT`, `ErrorWithCode.EXECUTION_ERROR`, `RawReply.EXECUTION_ERROR`
+- Added `DateTimeUtils.tryStringToDate`
 - Added `BeijingConv.tryStringToMillis`, `BeijingConv.tryStringToSeconds`, `BeijingConv.tryStringToOffsetDateTime`
 - Mark many scala constant to `final` to suggest compiler make inline optimization